refactor(sync): add mod client
