+++ crates/papyrus_node/src/config/pointers.rs
@@ -25,7 +25,7 @@ use papyrus_config::ParamPrivacyInput;
 use papyrus_config::{ConfigError, ParamPath, SerializedParam};
 use papyrus_monitoring_gateway::MonitoringGatewayConfig;
 use papyrus_network::NetworkConfig;
-use papyrus_p2p_sync::{P2PSync, P2PSyncConfig};
+use papyrus_p2p_sync::client::{P2PSync, P2PSyncConfig};
 #[cfg(feature = "rpc")]
 use papyrus_rpc::RpcConfig;
 use papyrus_storage::db::DbConfig;
